refactor(assistant): move conversation-starter helpers out of memory/ into home/
