+++ assistant/src/home/conversation-starter-validation.ts
