+++ assistant/src/memory/job-handlers/conversation-starters.ts
@@ -9,6 +9,21 @@ import { and, desc, eq, sql } from "drizzle-orm";
 import { v4 as uuid } from "uuid";
 
 import { loadSkillCatalog } from "../../config/skills.js";
+import {
+  checkpointKey,
+  CK_BATCH,
+  CK_ITEM_COUNT,
+  CK_LAST_ATTEMPT_AT,
+  CK_LAST_GEN_AT,
+  countActiveMemoryNodes,
+  getCheckpointValue,
+  parseCheckpointInt,
+  upsertCheckpoint,
+} from "../../home/conversation-starter-checkpoints.js";
+import {
+  buildConversationStarterValidationContext,
+  isValidConversationStarterText,
+} from "../../home/conversation-starter-validation.js";
 import { getDb } from "../../persistence/db-connection.js";
 import { asString } from "../../persistence/job-utils.js";
 import type { MemoryJob } from "../../persistence/jobs-store.js";
@@ -25,21 +40,6 @@ import {
 } from "../../providers/provider-send-message.js";
 import { getLogger } from "../../util/logger.js";
 import { truncate } from "../../util/truncate.js";
-import {
-  checkpointKey,
-  CK_BATCH,
-  CK_ITEM_COUNT,
-  CK_LAST_ATTEMPT_AT,
-  CK_LAST_GEN_AT,
-  countActiveMemoryNodes,
-  getCheckpointValue,
-  parseCheckpointInt,
-  upsertCheckpoint,
-} from "../conversation-starter-checkpoints.js";
-import {
-  buildConversationStarterValidationContext,
-  isValidConversationStarterText,
-} from "../conversation-starter-validation.js";
 import { rawAll } from "../raw-query.js";
 
 const log = getLogger("conversation-starters-gen");
